+++ vllm/v1/sample/rejection_sampler.py
@@ -18,6 +18,12 @@ from vllm.v1.sample.ops.penalties import apply_all_penalties
 from vllm.v1.sample.ops.topk_topp_sampler import apply_top_k_top_p
 from vllm.v1.sample.sampler import Sampler
 from vllm.v1.spec_decode.metadata import SpecDecodeMetadata
+from vllm.v1.spec_decode.spec_decode_pytorch_utils import (
+    expand_pytorch,
+    rejection_greedy_sample_pytorch,
+    rejection_random_sample_pytorch,
+    sample_recovered_tokens_pytorch,
+)
 
 logger = init_logger(__name__)
 
@@ -691,229 +697,6 @@ def sample_recovered_tokens(
             no_draft_probs=draft_probs is None,
         )
     return recovered_token_ids
-
-
-# ============================================================================
-# PyTorch CPU Fallback Implementations
-# ============================================================================
-
-
-def rejection_greedy_sample_pytorch(
-    output_token_ids: torch.Tensor,  # [batch_size, max_spec_len + 1]
-    cu_num_draft_tokens: torch.Tensor,  # [batch_size]
-    draft_token_ids: torch.Tensor,  # [num_tokens]
-    target_argmax: torch.Tensor,  # [num_tokens]
-    bonus_token_ids: torch.Tensor,  # [batch_size]
-    is_greedy: torch.Tensor | None,  # [batch_size] or None
-):
-    """PyTorch implementation of rejection_greedy_sample_kernel."""
-    batch_size = cu_num_draft_tokens.shape[0]
-    max_spec_len = output_token_ids.shape[1] - 1
-    device = draft_token_ids.device
-
-    # Compute counts and start indices per request
-    zero = torch.zeros(1, dtype=cu_num_draft_tokens.dtype, device=device)
-    counts = torch.diff(cu_num_draft_tokens, prepend=zero)
-    start_indices = torch.cat([zero, cu_num_draft_tokens[:-1]])
-
-    # Create request indices for each token
-    req_indices = torch.repeat_interleave(
-        torch.arange(batch_size, device=device), counts
-    )
-
-    # Create position indices within each request
-    pos_in_req = (
-        torch.arange(draft_token_ids.shape[0], device=device)
-        - start_indices[req_indices]
-    )
-
-    # Find mismatches (rejections)
-    is_mismatch = draft_token_ids != target_argmax
-
-    # For each request, find the first rejection position
-    # Set mismatches to their position, matches to max_spec_len + 1
-    mismatch_pos = torch.where(is_mismatch, pos_in_req, max_spec_len + 1)
-
-    # Scatter min to find first rejection per request
-    first_reject_pos = torch.full(
-        (batch_size,), max_spec_len + 1, dtype=torch.int64, device=device
-    )
-    first_reject_pos.scatter_reduce_(
-        0, req_indices.long(), mismatch_pos.long(), reduce="amin"
-    )
-
-    # Create output mask: positions < first_reject_pos get target_argmax
-    pos_range = torch.arange(max_spec_len, device=device).unsqueeze(0)
-    valid_mask = pos_range < first_reject_pos.unsqueeze(1)
-    valid_mask = valid_mask & (pos_range < counts.unsqueeze(1))
-
-    # Apply greedy mask if provided
-    if is_greedy is not None:
-        valid_mask = valid_mask & is_greedy.unsqueeze(1)
-
-    # Scatter target_argmax values to output
-    for req_idx in range(batch_size):
-        if is_greedy is not None and not is_greedy[req_idx]:
-            continue
-        n = counts[req_idx].item()
-        start = start_indices[req_idx].long().item()
-        first_rej = first_reject_pos[req_idx].item()
-
-        # Copy tokens up to and including the first rejection
-        copy_len = min(n, first_rej + 1)
-        output_token_ids[req_idx, :copy_len] = target_argmax[start : start + copy_len]
-
-        # Add bonus token if all accepted
-        if first_rej > n - 1:
-            output_token_ids[req_idx, n] = bonus_token_ids[req_idx, 0]
-
-
-def rejection_random_sample_pytorch(
-    output_token_ids: torch.Tensor,  # [batch_size, max_spec_len + 1]
-    cu_num_draft_tokens: torch.Tensor,  # [batch_size]
-    draft_token_ids: torch.Tensor,  # [num_tokens]
-    draft_probs: torch.Tensor | None,  # [num_tokens, vocab_size] or None
-    target_probs: torch.Tensor,  # [num_tokens, vocab_size]
-    bonus_token_ids: torch.Tensor,  # [batch_size]
-    recovered_token_ids: torch.Tensor,  # [num_tokens]
-    uniform_probs: torch.Tensor,  # [num_tokens]
-    is_greedy: torch.Tensor | None,  # [batch_size] or None
-    no_draft_probs: bool = False,
-):
-    """PyTorch implementation of rejection_random_sample_kernel."""
-    if not no_draft_probs and draft_probs is None:
-        raise ValueError("draft_probs is required when no_draft_probs=False")
-
-    batch_size = cu_num_draft_tokens.shape[0]
-    num_tokens = draft_token_ids.shape[0]
-    max_spec_len = output_token_ids.shape[1] - 1
-    device = draft_token_ids.device
-
-    # Compute counts and start indices per request
-    zero = torch.zeros(1, dtype=cu_num_draft_tokens.dtype, device=device)
-    counts = torch.diff(cu_num_draft_tokens, prepend=zero)
-    start_indices = torch.cat([zero, cu_num_draft_tokens[:-1]])
-
-    # Create request indices for each token
-    req_indices = torch.repeat_interleave(
-        torch.arange(batch_size, device=device), counts
-    )[:num_tokens]
-
-    # Create position indices within each request
-    pos_in_req = (
-        torch.arange(num_tokens, device=device) - start_indices[req_indices].long()
-    )
-
-    # Gather draft and target probabilities for the draft tokens
-    token_range = torch.arange(num_tokens, device=device)
-    if no_draft_probs:
-        draft_prob_vals = torch.ones(num_tokens, device=device)
-    else:
-        draft_prob_vals = draft_probs[token_range, draft_token_ids]  # type: ignore
-
-    target_prob_vals = target_probs[token_range, draft_token_ids]
-
-    # Compute acceptance: target_prob / draft_prob >= uniform_prob
-    # Handle division by zero
-    accept_ratio = torch.where(
-        draft_prob_vals > 0,
-        target_prob_vals / draft_prob_vals,
-        torch.zeros_like(target_prob_vals),
-    )
-    is_accepted = accept_ratio >= uniform_probs
-
-    # Find first rejection position per request
-    reject_pos = torch.where(is_accepted, max_spec_len + 1, pos_in_req)
-    first_reject_pos = torch.full(
-        (batch_size,), max_spec_len + 1, dtype=torch.int64, device=device
-    )
-    first_reject_pos.scatter_reduce_(
-        0, req_indices.long(), reject_pos.long(), reduce="amin"
-    )
-
-    # Process each request
-    for req_idx in range(batch_size):
-        if is_greedy is not None and is_greedy[req_idx]:
-            continue
-
-        n = counts[req_idx].item()
-        start = int(start_indices[req_idx].item())
-        first_rej = first_reject_pos[req_idx].item()
-
-        # Copy accepted tokens (draft_token_ids) and first rejected (recovered)
-        for pos in range(min(n, first_rej)):
-            output_token_ids[req_idx, pos] = draft_token_ids[start + pos]
-
-        # If there was a rejection, add the recovered token
-        if first_rej < n:
-            output_token_ids[req_idx, first_rej] = recovered_token_ids[
-                start + first_rej
-            ]
-
-        # Add bonus token if all accepted
-        if first_rej >= n:
-            output_token_ids[req_idx, n] = bonus_token_ids[req_idx, 0]
-
-
-def expand_pytorch(
-    output: torch.Tensor,  # [num_tokens]
-    input_val: torch.Tensor,  # [batch_size]
-    cu_num_tokens: torch.Tensor,  # [batch_size]
-    replace_from: int,
-    replace_to: int,
-):
-    """PyTorch implementation of expand_kernel."""
-    # Compute counts per batch from cumulative sum
-    zero = torch.zeros(1, dtype=cu_num_tokens.dtype, device=cu_num_tokens.device)
-    counts = torch.diff(cu_num_tokens, prepend=zero)
-
-    # Replace values
-    vals = torch.where(input_val == replace_from, replace_to, input_val)
-
-    # Expand using repeat_interleave
-    total_tokens = cu_num_tokens[-1].item() if len(cu_num_tokens) > 0 else 0
-    output[:total_tokens] = vals.repeat_interleave(counts).to(output.dtype)
-
-
-def sample_recovered_tokens_pytorch(
-    output_token_ids: torch.Tensor,  # [num_tokens]
-    cu_num_draft_tokens: torch.Tensor,  # [batch_size]
-    draft_token_ids: torch.Tensor,  # [num_tokens]
-    draft_probs: torch.Tensor | None,  # [num_tokens, vocab_size] or None
-    target_probs: torch.Tensor,  # [num_tokens, vocab_size]
-    inv_q: torch.Tensor,  # [batch_size, vocab_size] - reciprocal of q
-    no_draft_probs: bool = False,
-):
-    """PyTorch implementation of sample_recovered_tokens_kernel."""
-    if not no_draft_probs and draft_probs is None:
-        raise ValueError("draft_probs is required when no_draft_probs=False")
-
-    num_tokens = draft_token_ids.shape[0]
-    device = target_probs.device
-
-    # Map each token position to its request index
-    zero = torch.zeros(1, dtype=cu_num_draft_tokens.dtype, device=device)
-    counts = torch.diff(cu_num_draft_tokens, prepend=zero)
-    req_indices = torch.repeat_interleave(
-        torch.arange(len(counts), device=device), counts
-    )[:num_tokens]
-
-    # Gather inv_q values for each token position
-    inv_q_expanded = inv_q[req_indices]  # [num_tokens, vocab_size]
-
-    if no_draft_probs:
-        # Zero out the draft token probability for each position
-        prob = target_probs.clone()
-        token_indices = torch.arange(num_tokens, device=device)
-        prob[token_indices, draft_token_ids] = 0
-    else:
-        # Compute adjusted probability: max(target - draft, 0)
-        diff = target_probs - draft_probs  # type: ignore[operator]
-        prob = torch.clamp(diff, min=0)
-
-    # Gumbel-max trick: argmax(prob * inv_q) equivalent to argmax(prob / q)
-    recovered_ids = torch.argmax(prob * inv_q_expanded, dim=1)
-    output_token_ids[:num_tokens] = recovered_ids.to(output_token_ids.dtype)
 
 
 # NOTE(woosuk): Avoid specialization to prevent unnecessary recompilation.
